Refactor API into separate folder, remove api.js, document API functions
Conflicts:
	packages/perspective/src/js/API/dispatch.js
	packages/perspective/src/js/api.js
